fix update method of profiles controller expecting user param.

+++ app/controllers/profiles_controller.rb
@@ -14,6 +14,7 @@ class ProfilesController < ApplicationController
   end
 
   def update
+    # Handle equipment updates
     if profile_params[:equipment_ids].present?
       # Clear existing equipment associations
       current_user.user_equipments.destroy_all
@@ -22,11 +23,18 @@ class ProfilesController < ApplicationController
       profile_params[:equipment_ids].each do |equipment_id|
         current_user.user_equipments.create(equipment_id: equipment_id)
       end
-    else
+    elsif params[:profile].present? && !profile_params[:equipment_ids].present?
+      # If profile is submitted but no equipment IDs, clear all equipment
       current_user.user_equipments.destroy_all
     end
 
-    if @user.update(user_params)
+    # Handle user details update if provided
+    user_updated = true
+    if params[:user].present?
+      user_updated = @user.update(user_params)
+    end
+
+    if user_updated
       redirect_to profile_path, notice: 'Profile updated successfully'
     else
       @equipment_by_category = Equipment.all.group_by(&:category)
@@ -42,10 +50,10 @@ class ProfilesController < ApplicationController
   end
 
   def profile_params
-    params.require(:profile).permit(equipment_ids: [])
+    params.fetch(:profile, {}).permit(equipment_ids: [])
   end
 
   def user_params
-    params.require(:user).permit(:email, :name)
+    params.require(:user).permit(:name) if params[:user].present?
   end
 end
